implement scrolling feature in feed and item lists, raise maximum length for indexes to 20 characters, use windows instead of pads in ncurses, now item index is of int64_t type, raise default max items to 100

+++ src/config.h
@@ -2,6 +2,6 @@
 #define CONFIG_H
 extern uint8_t config_top_offset;
 extern uint8_t config_left_offset;
-extern uint16_t config_max_items;
+extern int64_t config_max_items;
 extern uint8_t config_number;
 #endif // CONFIG_H
